requirements.txt
boto3==1.12.14
botocore==1.15.14
pandas==1.1.3
pyarrow==1.0.1
gcsfs==0.7.1
google-cloud-bigquery-storage==2.0.0
pycountry==20.7.3
s3transfer==0.3.3
python-dotenv==0.14.0
